refactor: restructure interface
