+++ inlang/source-code/website/src/interface/README.md
@@ -1,4 +1,11 @@
-UI components used across the app.
+# inlang Interface for website
+UI components, sections and helper functions used across the app.
+
+This folder contains:
+- [Components](/components)
+- [Sections](/sections)
+- [Icons](/custom-icons)
+... and some smaller files like helpers.
 
 ### Using dynamic colors like `bg-${props.color}`
