+++ inlang/source-code/website/src/interface/custom-icons/pluginIcon.tsx
