+++ inlang/source-code/website/src/interface/editor/EditorLayout.tsx
