+++ inlang/source-code/website/src/interface/sections/marketplace/index.tsx
@@ -1,7 +1,7 @@
 import { For, Show, createSignal, Match, Switch, onMount } from "solid-js"
 // import { SearchIcon } from "#src/pages/editor/@host/@owner/@repository/components/SearchInput.jsx"
 // import { Button } from "#src/pages/index/components/Button.jsx"
-import { GetHelp } from "#src/components/GetHelp.jsx"
+import { GetHelp } from "#src/interface/components/GetHelp.jsx"
 // import Check from "~icons/material-symbols/check"
 import Right from "~icons/material-symbols/chevron-right"
 import Left from "~icons/material-symbols/chevron-left"
@@ -10,8 +10,8 @@ import { currentPageContext } from "#src/renderer/state.js"
 // @ts-ignore
 import { createSlider } from "solid-slider"
 import "solid-slider/slider.css"
-import Highlight from "#src/components/Highlight.jsx"
-import Card, { CardBuildOwn, NoResultsCard } from "#src/components/Card.jsx"
+import Highlight from "#src/interface/components/Highlight.jsx"
+import Card, { CardBuildOwn, NoResultsCard } from "#src/interface/components/Card.jsx"
 import { rpc } from "@inlang/rpc"
 import { registry } from "@inlang/marketplace-registry"
