+++ inlang/source-code/website/src/pages/c/@category/index.page.tsx
@@ -1,11 +1,11 @@
 import { For, Show, createSignal, onMount } from "solid-js"
-import { GetHelp } from "#src/components/GetHelp.jsx"
+import { GetHelp } from "#src/interface/components/GetHelp.jsx"
 import { SectionLayout } from "#src/pages/index/components/sectionLayout.jsx"
 import { currentPageContext } from "#src/renderer/state.js"
-import Highlight from "#src/components/Highlight.jsx"
-import Card, { CardBuildOwn, NoResultsCard } from "#src/components/Card.jsx"
+import Highlight from "#src/interface/components/Highlight.jsx"
+import Card, { CardBuildOwn, NoResultsCard } from "#src/interface/components/Card.jsx"
 import { Meta, Title } from "@solidjs/meta"
-import MarketplaceLayout from "#src/components/marketplace/MarketplaceLayout.jsx"
+import MarketplaceLayout from "#src/interface/marketplace/MarketplaceLayout.jsx"
 
 type SubCategoryApplication = "app" | "library" | "plugin" | "messageLintRule"
