+++ inlang/source-code/website/src/pages/index/custom_section/Hero.tsx
@@ -1,6 +1,6 @@
 import { For, Show } from "solid-js"
 import { registry } from "@inlang/marketplace-registry"
-import { Chip } from "#src/components/Chip.jsx"
+import { Chip } from "#src/interface/components/Chip.jsx"
 import { colorForTypeOf, typeOfIdToTitle } from "#src/pages/m/utilities.js"
 import { Button } from "../components/Button.jsx"
